feat(core): guard all core update entries against rc.2 breaking changes
Centralize the above-rc.2 warning into a single useCoreBreakingConfirm hook (hardcoded 0.1.1-rc.2 baseline, independent of recommend-version logic) and call it from every core-version update entry point: the core panel manual download (config-core), the core '新版本' toast trigger (harness-updater), and the debug panel '存在新版本' link (config-debug). Cancelling the warning aborts the download; confirming proceeds. Shared compareVersions moved to utils/core-version.ts.
