+++ src/components/config-core.tsx
@@ -6,7 +6,9 @@ import { invoke } from '@tauri-apps/api/core'
 import { Fragment, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { If } from 'react-if-lite'
+import { useCoreBreakingConfirm } from '@/hooks/use-core-breaking-confirm'
 import { store } from '@/store'
+import { compareVersions } from '@/utils/core-version'
 import { toast } from '@/utils/toast'
 import { useDshCores } from '../hooks/use-dsh-cores'
 import { DownloadCoreDialog } from './download-core-dialog'
@@ -32,19 +34,10 @@ import { PanelState } from './panel-state'
  * - 每行展示核心入口（cli path，超长省略号 + 限制宽度）。
  */
 
-/** rc.2 基准版本：高于该版本的核心引入破坏性更改，可能影响第三方插件 */
-const RC2_BASELINE = '0.1.1-rc.2'
-
-/** 判断核心版本是否高于 rc.2 基准（引入破坏性更改） */
-function isAboveRc2(version: string): boolean {
-  if (!version)
-    return false
-  return compareVersions(version, RC2_BASELINE) > 0
-}
-
 export function ConfigCore() {
   const [dialogHolder, openDialog] = useOverlay(Modal, { type: 'holder' })
   const [downloadDialogHolder, openDownloadDialog] = useOverlay(DownloadCoreDialog, { type: 'holder' })
+  const { holder: coreBreakingHolder, confirmCoreBreaking } = useCoreBreakingConfirm()
 
   const { t } = useTranslation()
   const { cores, loading, error, setActiveCore, updateLocalCore, downloadCore, removeCore, refreshCores, busy } = useDshCores()
@@ -143,24 +136,9 @@ export function ConfigCore() {
     if (busy)
       return
     // rc.2 以上版本引入破坏性更改，可能影响第三方插件 → 下载前先弹出确认。
-    // 该提示固定以 rc.2 为基准（与推荐版本逻辑无关），仅用于告知用户，
-    // 并提示可随时在核心面板切回 rc.2。
-    if (isAboveRc2(core.version)) {
-      try {
-        await openDialog({
-          status: 'warning',
-          title: t('core.above_rc2_warning_title'),
-          description: (
-            <p>
-              {t('core.above_rc2_warning_desc', { version: RC2_BASELINE })}
-            </p>
-          ),
-        })
-      }
-      catch {
-        return
-      }
-    }
+    // 取消则中止，确认后继续；该提示与推荐版本逻辑无关，仅告知用户可随时切回 rc.2。
+    if (!(await confirmCoreBreaking(core.version)))
+      return
     // 下载过程在对话框内展示进度 + 日志（复用 install-progress 事件流）；
     // 对话框 confirm（下载成功）或 cancel（失败后点关闭）都会结束本次等待。
     try {
@@ -412,6 +390,7 @@ export function ConfigCore() {
 
       {dialogHolder}
       {downloadDialogHolder}
+      {coreBreakingHolder}
     </div>
   )
 }
@@ -419,61 +398,4 @@ export function ConfigCore() {
 /** 版本展示：优先版本号，缺失回落来源 id */
 function displayVersion(version: HarnessCore): string {
   return version.version || (version.source === 'local' ? 'local' : 'app')
-}
-
-/**
- * 简化 semver 比较（不引入额外依赖），可处理 `0.1.1-rc.2` / `0.1.1` 格式。
- * 返回值：负数 a < b，0 相等，正数 a > b。
- */
-function compareVersions(a: string, b: string): number {
-  const parse = (value: string) => {
-    let v = value
-    while (v.startsWith('dsh-') || v.startsWith('src-'))
-      v = v.replace(/^(?:src|dsh)-/, '')
-    if (v.includes('-')) {
-      const candidate = v.slice(0, v.lastIndexOf('-'))
-      if (/^\d+\.\d+\.\d+(?:-[0-9A-Z.-]+)?$/i.test(candidate))
-        v = candidate
-    }
-    const [core, pre = ''] = v.split('-', 2)
-    const nums = core.split('.').map(n => parseInt(n, 10) || 0)
-    return { nums, pre }
-  }
-  const pa = parse(a)
-  const pb = parse(b)
-  for (let i = 0; i < 3; i++) {
-    const x = pa.nums[i] ?? 0
-    const y = pb.nums[i] ?? 0
-    if (x !== y)
-      return x < y ? -1 : 1
-  }
-  // 无预发布号 > 有预发布号
-  if (!pa.pre && !pb.pre)
-    return 0
-  if (!pa.pre)
-    return 1
-  if (!pb.pre)
-    return -1
-  // 预发布号按点分段比较：数字按数值、非数字按字典序
-  const paParts = pa.pre.split('.').map(p => (Number.isNaN(Number(p)) ? p : Number(p)))
-  const pbParts = pb.pre.split('.').map(p => (Number.isNaN(Number(p)) ? p : Number(p)))
-  const len = Math.max(paParts.length, pbParts.length)
-  for (let i = 0; i < len; i++) {
-    const x = paParts[i]
-    const y = pbParts[i]
-    if (x === undefined)
-      return -1
-    if (y === undefined)
-      return 1
-    if (x === y)
-      continue
-    if (typeof x === 'number' && typeof y === 'number')
-      return x < y ? -1 : 1
-    if (typeof x === 'number')
-      return -1
-    if (typeof y === 'number')
-      return 1
-    return x < y ? -1 : 1
-  }
-  return 0
 }
